Move Mock Classes into separate Files in CriticalMapsMocksForTests Directory
