+++ CriticalMassTests/XCTestCase+Helper.swift
@@ -1,0 +1,24 @@
+//
+//  XCTestCase+Helper.swift
+//  CriticalMaps
+//
+//  Created by Leonard Thomas on 5/12/19.
+//  Copyright © 2019 Pokus Labs. All rights reserved.
+//
+
+@testable import CriticalMaps
+import XCTest
+
+extension XCTestCase {
+    func wait(interval: TimeInterval, completion: @escaping () -> Void) {
+        Timer.scheduledTimer(withTimeInterval: interval, repeats: false) { _ in
+            completion()
+        }
+    }
+
+    func execute(times: UInt, _ function: @autoclosure () -> Void) {
+        (0 ..< times).forEach { _ in
+            function()
+        }
+    }
+}
